Update AlphaPrime package and add new components
Updated the version of AlphaPrime from 15.0.1 to 15.0.2 and added a dependency on "@fortawesome/fontawesome-free" (version 6.5.2). Introduced two new components: AlphaAddButtonComponent and AlphaDebugTagComponent. These components, along with the new AlphaPrimeService, are now exposed in the package's public API. Additionally, a deprecated notice was added to the dialog service initialisation method in AlphaPrimeModalService.