Upgrade versions and implement additional services and components
This commit includes upgrading alpha-prime version to 15.0.13 and alpha-api version to 17.0.4. It also includes the implementation of additional components and services such as alpha-prime-filter-box and alpha-version-api.service. These changes have been made respectively in both A15 and A17 projects. The corresponding testing modules have also been added to check the functionality of the new implementations.