Refactor alpha-prime service and component
The alpha-prime service has seen its structure refactored for better readability and usage, introducing UploadApiService and LocalBusService. The changes in the service also required an update in the alpha-prime component, mainly in the methods of upload and delete. Furthermore, AlphaPrimeLabelComponent, a new visual element for the user interface, has been added. The version got upgraded to 15.0.15 to reflect these changes. The improvements aim to optimize the flow and usage of the project for further development.