Add Alpha Prime Number Input Component and update project version
Introduced a new component 'Alpha Prime Number Input' in the Alpha Prime project. In addition to the creation of all necessary files for this component, the version of the project has been updated to 15.0.18. This new component enhances the input management system and supports various numeric input scenarios, thereby improving the overall usability of the project.