Add new interfaces and update existing ones in alpha-com project
The update includes the addition of several new interfaces such as `IAlphaPage`, `IAlphaPrincipal`, `IAlphaTranslationItem`, and more. The method signatures of certain interfaces like `format` and `formatRange` in `alpha-ymd.ts` were also updated to allow an additional parameter. The project version was updated to `15.0.3`. These changes aid in improving code readability and maintainability.