Refactor to utilize @pvway/alpha-common and update service implementations
This commit involves removing some redundant interfaces and shifting their references to the @pvway/alpha-common library. A significant change is updating AlphaOasService to implement IAlphaOAuthService present in the shared library. It also shows updating package.json to include @pvway/alpha-common as a new peer dependency and reflects the version bumps in package-lock.json.