Refactor alpha-ems to initialize HttpClient in service
The HttpClient used in AlphaEmsBaseApi is now initialized directly in AlphaEmsService, leading to reduction of required parameters in different functions and constructors of AlphaEmsBaseApi. The AlphaEmsBaseApi now takes ems service as a single parameter in constructors. This change aims to improve the modularity and isolation of services. Additionally, project version bumped from 17.3.4 to 17.4.0.