Add a new library for translations
A library named AlphaTs has been added which contains a translation service. This service helps to manage all client-side translations. It communicates with a server endpoint to check and fetch updated translations if needed. The library also includes various related features and services to support translations and facilitate its usage.