Add AlphaPrime service and components
This commit includes the addition of the AlphaPrime service that provides functionalities for handling modals in the application. Several new directives, components, services and tests have also been added along with also the jest setup file for testing purposes. Changes are also made to the config files to include the new service and components.