Update version and add guard function in alpha-ns service
The version of alpha-ns in package.json has been updated to 15.2.1, and the description was changed to better reflect library purpose. In the alpha-ns.service.ts, unnecessary whitespace was removed. More importantly, it's added a new 'guard' function, responsible for conditional navigation, providing high flexibility regarding redirections. Also, a small format fix was applied to the assignment of the _postNavigationLog variable.