Add redirection guard method to alpha service
A new method, `guard`, has been added to the alpha service. This method redirects to a specified page if a certain condition, provided as a function, is true. Logic has also been added to test this new method. The description of the alpha navigation service has also been updated in package.json.